[Feature] Move GDAX to CoinbasePro
